+++ crates/miden-objects/src/transaction/mod.rs
@@ -6,24 +6,24 @@ use super::{
     vm::AdviceInputs,
 };
 
-mod chain_mmr;
 mod executed_tx;
 mod inputs;
 mod ordered_transactions;
 mod outputs;
+mod partial_blockchain;
 mod proven_tx;
 mod transaction_id;
 mod tx_args;
 mod tx_header;
 mod tx_witness;
 
-pub use chain_mmr::ChainMmr;
 pub use executed_tx::{ExecutedTransaction, TransactionMeasurements};
 pub use inputs::{
     ForeignAccountInputs, InputNote, InputNotes, ToInputNoteCommitments, TransactionInputs,
 };
 pub use ordered_transactions::OrderedTransactionHeaders;
 pub use outputs::{OutputNote, OutputNotes, TransactionOutputs};
+pub use partial_blockchain::PartialBlockchain;
 pub use proven_tx::{
     InputNoteCommitment, ProvenTransaction, ProvenTransactionBuilder, TxAccountUpdate,
 };